requirements.txt
# Core
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Financial Data
yfinance>=0.2.36
pandas-datareader>=0.10.0

# ML Models
lightgbm>=4.0.0
xgboost>=2.0.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Technical Analysis
ta>=0.11.0

# LLM (optional - system works without these)
groq>=0.4.0
openai>=1.0.0
anthropic>=0.20.0

# Export
jinja2>=3.1.0
